examples/libraries/requirements.txt
opengris-parfun[scaler]
pargraph
